+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+# 9999
+
+* adding `postCompileCheckLog` to make error log an option
+
 # 0.1.0.0
 
 * initial version